PYIC-7650 Move fraud availability check entirely into VcHelper
