+++ libs/user-identity-service/src/test/java/uk/gov/di/ipv/core/library/service/UserIdentityServiceTest.java
@@ -123,73 +123,6 @@ class UserIdentityServiceTest {
     }
 
     @Test
-    void shouldAllowM1cForApplicableAuthoritativeSourceFailedFraudCheck() {
-
-        // Arrange
-        var vcs =
-                List.of(
-                        DCMAW_PASSPORT_VC,
-                        M1A_ADDRESS_VC,
-                        vcFraudApplicableAuthoritativeSourceFailed(),
-                        vcVerificationM1a());
-
-        // Act
-        var result = userIdentityService.isFraudCheckUnavailable(vcs);
-
-        // Assert
-        assertTrue(result);
-    }
-
-    @Test
-    void shouldNotAllowM1cForOtherFailedFraudCheck() {
-
-        // Arrange
-        var vcs =
-                List.of(
-                        DCMAW_PASSPORT_VC,
-                        M1A_ADDRESS_VC,
-                        vcExperianFraudEvidenceFailed(),
-                        vcVerificationM1a());
-
-        // Act
-        var result = userIdentityService.isFraudCheckUnavailable(vcs);
-
-        // Assert
-        assertFalse(result);
-    }
-
-    @Test
-    void shouldNotAllowM1cForSuccessfulFraudCheck() {
-
-        // Arrange
-        var vcs =
-                List.of(
-                        DCMAW_PASSPORT_VC,
-                        M1A_ADDRESS_VC,
-                        vcExperianFraudScoreOne(),
-                        vcVerificationM1a());
-
-        // Act
-        var result = userIdentityService.isFraudCheckUnavailable(vcs);
-
-        // Assert
-        assertFalse(result);
-    }
-
-    @Test
-    void shouldNotAllowM1cForMissingFraudCheck() {
-
-        // Arrange
-        var vcs = List.of(DCMAW_PASSPORT_VC, M1A_ADDRESS_VC, vcVerificationM1a());
-
-        // Act
-        var result = userIdentityService.isFraudCheckUnavailable(vcs);
-
-        // Assert
-        assertFalse(result);
-    }
-
-    @Test
     void shouldSetVotClaimToP2OnSuccessfulIdentityCheck() throws Exception {
         // Arrange
         var vcs = List.of(PASSPORT_NON_DCMAW_SUCCESSFUL_VC, vcExperianFraudScoreTwo(), VC_ADDRESS);
