+++ libs/user-identity-service/src/test/java/uk/gov/di/ipv/core/library/service/VotMatcherTest.java
@@ -25,6 +25,7 @@ import static uk.gov.di.ipv.core.library.enums.Vot.PCL200;
 import static uk.gov.di.ipv.core.library.enums.Vot.PCL250;
 import static uk.gov.di.ipv.core.library.enums.Vot.SUPPORTED_VOTS_BY_DESCENDING_STRENGTH;
 import static uk.gov.di.ipv.core.library.fixtures.VcFixtures.vcExperianFraudScoreTwo;
+import static uk.gov.di.ipv.core.library.fixtures.VcFixtures.vcFraudApplicableAuthoritativeSourceFailed;
 import static uk.gov.di.ipv.core.library.fixtures.VcFixtures.vcHmrcMigrationPCL200;
 import static uk.gov.di.ipv.core.library.fixtures.VcFixtures.vcHmrcMigrationPCL250;
 import static uk.gov.di.ipv.core.library.fixtures.VcFixtures.vcVerificationM1a;
@@ -143,18 +144,18 @@ public class VotMatcherTest {
     @Test
     void shouldMatchM1cIfFraudCheckUnavailable() throws Exception {
         // Arrange
+        var vcs = List.of(vcVerificationM1a(), vcFraudApplicableAuthoritativeSourceFailed());
         var expectedProfiles =
                 List.of(Gpg45Profile.M1A, Gpg45Profile.M1B, Gpg45Profile.M2B, Gpg45Profile.M1C);
-        when(mockUseridentityService.checkRequiresAdditionalEvidence(gpg45Vcs)).thenReturn(false);
-        when(mockUseridentityService.isFraudCheckUnavailable(gpg45Vcs)).thenReturn(true);
-        when(mockGpg45ProfileEvaluator.buildScore(gpg45Vcs)).thenReturn(GPG_45_SCORES);
+        when(mockUseridentityService.checkRequiresAdditionalEvidence(vcs)).thenReturn(false);
+        when(mockGpg45ProfileEvaluator.buildScore(vcs)).thenReturn(GPG_45_SCORES);
         when(mockGpg45ProfileEvaluator.getFirstMatchingProfile(GPG_45_SCORES, expectedProfiles))
                 .thenReturn(Optional.of(Gpg45Profile.M1C));
 
         // Act
         var votMatch =
                 votMatcher.matchFirstVot(
-                        SUPPORTED_VOTS_BY_DESCENDING_STRENGTH, gpg45Vcs, List.of(), true);
+                        SUPPORTED_VOTS_BY_DESCENDING_STRENGTH, vcs, List.of(), true);
 
         // Assert
         verify(mockGpg45ProfileEvaluator).getFirstMatchingProfile(GPG_45_SCORES, expectedProfiles);
@@ -167,7 +168,6 @@ public class VotMatcherTest {
         // Arrange
         var expectedProfiles = List.of(Gpg45Profile.M1A, Gpg45Profile.M1B, Gpg45Profile.M2B);
         when(mockUseridentityService.checkRequiresAdditionalEvidence(gpg45Vcs)).thenReturn(false);
-        when(mockUseridentityService.isFraudCheckUnavailable(gpg45Vcs)).thenReturn(false);
         when(mockGpg45ProfileEvaluator.buildScore(gpg45Vcs)).thenReturn(GPG_45_SCORES);
         when(mockGpg45ProfileEvaluator.getFirstMatchingProfile(GPG_45_SCORES, expectedProfiles))
                 .thenReturn(Optional.of(Gpg45Profile.M1C));
